@@ -11,7 +11,7 @@
     "test-visual": "percy exec -- ./node_modules/.bin/cypress run",
     "test-visual:open": "start-server-and-test fake http://localhost:9000/ 'percy exec -- ./node_modules/.bin/cypress open'",
     "test-visual:wait": "percy build:wait -f --project f10a8acf/boclips-web-app --commit HEAD",
-    "dev": "NODE_ENV=development TAILWIND_MODE=watch vite serve --config vite/vite.dev.js -d",
+    "dev": "NODE_ENV=development TAILWIND_MODE=watch vite serve --config vite-dev.config.js -d --force",
     "local": "NODE_ENV=development TAILWIND_MODE=watch vite serve --config webpack/webpack.local.js",
     "fake": "NODE_ENV=development TAILWIND_MODE=watch vite serve --config webpack/webpack.fake.js",
     "compile": "tsc -p . --noEmit",
@@ -33,7 +33,7 @@
   },
   "homepage": "https://github.com/boclips/boclips-web-app#readme",
   "engines": {
-    "node": ">=20.17.0 <21"
+    "node": ">=20.18.0 <21"
   },
   "devDependencies": {
     "@percy/cli": "^1.30.2",
@@ -83,9 +83,6 @@
     "tailwindcss": "^3.4.15",
     "tailwindcss-debug-screens": "^2.2.1",
     "type-fest": "^4.18.2",
-    "vite-plugin-html": "^3.2.2",
-    "vite-plugin-html-purgecss": "^0.1.1",
-    "vite-plugin-svgr": "^4.3.0",
     "webpack-bundle-analyzer": "^4.10.2",
     "webpack-dev-server": "^4.15.1",
     "webpack-merge": "^5.10.0"
@@ -132,9 +129,9 @@
     "axios": "^0.28.1",
     "babel-loader": "^9.2.1",
     "boclips-js-security": "^13.0.1",
-    "boclips-ui": "^1.0.0",
     "boclips-player": "^7.16.0",
     "boclips-player-react": "^4.0.7",
+    "boclips-ui": "^1.0.0",
     "copy-webpack-plugin": "^11.0.0",
     "core-js": "^3.39.0",
     "dateformat": "^4.4.1",
@@ -154,9 +151,9 @@
     "querystring-es3": "^0.2.1",
     "rc-checkbox": "^3.3.0",
     "react": "18.3.1",
-    "react-dom": "18.3.1",
     "react-copy-to-clipboard": "5.1.0",
     "react-country-flag": "^3.1.0",
+    "react-dom": "18.3.1",
     "react-google-recaptcha-v3": "^1.10.1",
     "react-helmet": "^6.1.0",
     "react-password-validattor": "^1.2.0",
@@ -169,6 +166,8 @@
     "tsparticles": "^3.5.0",
     "typescript": "^5.6.3",
     "urijs": "^1.19.11",
-    "vite": "^6.0.2"
+    "vite": "^6.0.2",
+    "vite-plugin-html": "^3.2.2",
+    "vite-plugin-svgr": "^4.3.0"
   }
 }
