@@ -1,6 +1,5 @@
 import dayjs from 'dayjs';
-
-const duration = require('dayjs/plugin/duration');
+import duration from 'dayjs/plugin/duration';
 
 dayjs.extend(duration);
 
